fix: Refatoração do arquivo VindiWoocommerce.php com mais de 250 linhas. Métodos foram extraídos para melhorar a legibilidade e atender às recomendações do CodeClimate
