+++ src/assets/css/frontend.css
@@ -62,7 +62,7 @@
 .vindi_cc_card-container {
   width: 100%;
   max-width: 400px;
-  max-height: 180px;
+  max-height: 260px;
   height: 54vw;
 }
