+++ src/assets/js/admin.js
@@ -1,4 +1,4 @@
-jQuery(document).ready(function($) {
+jQuery(document).ready(function ($) {
   'use strict';
 
   var cycles_input = document.querySelector('#cycle_count'),
@@ -9,10 +9,10 @@ jQuery(document).ready(function($) {
     $interest_rate_input = $(interest_rate_input);
   var wcs_number_payments = document.querySelector('#wcs_number_payments'),
     $wcs_number_payments = $(wcs_number_payments);
-  
+
   $interest_rate_input.mask('##0.00%', {
     reverse: true,
-    onKeyPress: function(val, e, field, options) {
+    onKeyPress: function (val, e, field, options) {
       var old_value = $(field).data('oldValue') || '';
 
       val = val.trim();
@@ -35,9 +35,8 @@ jQuery(document).ready(function($) {
     }
   });
 
-  if ($wcs_number_payments)
-  {
-    $wcs_number_payments.on('change', function() {
+  if ($wcs_number_payments) {
+    $wcs_number_payments.on('change', function () {
       var val = $(this).val();
       var regex = new RegExp(/^(1|2|3|4|5|6|7|8|9|10|11|12)$/);
       if (val != '' && !regex.test(val)) {
@@ -45,7 +44,7 @@ jQuery(document).ready(function($) {
       }
     });
   }
-  
+
   /**
    * Subscription coupon actions.
    * @type {{init: function, type_options: function, move_field: function}}
@@ -55,7 +54,7 @@ jQuery(document).ready(function($) {
     /**
      * Initialize variation actions.
      */
-    init: function() {
+    init: function () {
       if (cycles_field) {
         $(document.getElementById('discount_type')).on('change', this.type_options).change();
         this.move_field();
@@ -65,27 +64,24 @@ jQuery(document).ready(function($) {
     /**
      * Show/hide fields by coupon type options.
      */
-    type_options: function() {
+    type_options: function () {
       var discount_type = $(this).val();
 
       switch (discount_type) {
         case 'recurring_percent':
-          if ($cycles_field.is(':visible'))
-          {
+          if ($cycles_field.is(':visible')) {
             $cycles_field.hide();
             $cycles_input.val('0');
           }
           break;
         case 'fixed_cart':
-          if ($cycles_field.is(':visible'))
-          {
+          if ($cycles_field.is(':visible')) {
             $cycles_field.hide();
             $cycles_input.val('1');
           }
           break;
         default:
-          if ($cycles_field.is(':hidden'))
-          {
+          if ($cycles_field.is(':hidden')) {
             $cycles_field.show();
             $cycles_input.val('0');
           }
@@ -96,7 +92,7 @@ jQuery(document).ready(function($) {
     /**
      * Move the renewal form field in the DOM to a better location.
      */
-    move_field: function() {
+    move_field: function () {
       var parent = document.getElementById('general_coupon_data'),
         shipping = parent.querySelector('.free_shipping_field');
 
@@ -112,7 +108,7 @@ jQuery(document).ready(function($) {
       e.preventDefault();
     });
   }
-  
+
   const btnCopyPost = document.querySelector('#buttonCopyPost');
   btnCopyPost.addEventListener("click", function (e) {
     e.preventDefault();
