+++ src/validators/WCCartSubscriptionLimiter.php
@@ -1,0 +1,137 @@
+<?php
+
+namespace VindiPaymentGateways;
+
+use WC_Subscriptions_Product;
+
+if (!defined('ABSPATH')) {
+    exit;
+}
+
+class WCCartSubscriptionLimiter
+{
+    public function __construct()
+    {
+        add_filter('woocommerce_add_to_cart_validation', [$this, 'limit_same_subscriptions'], 10, 3);
+        add_filter('woocommerce_update_cart_validation', [$this, 'limit_duplicate_subscriptions_cart_update'], 10, 4);
+        add_filter('woocommerce_add_to_cart_validation', [$this, 'disallow_subscription_single_product_cart'], 10, 4);
+    }
+
+    public function limit_same_subscriptions($passed, $product_id, $quantity)
+    {
+        $product = wc_get_product($product_id);
+
+        if ($product->is_virtual()) {
+            return $passed;
+        }
+
+        if (WC_Subscriptions_Product::is_subscription($product_id)) {
+            $subscription_count = $this->get_subscription_count($product_id);
+
+            if ($subscription_count + $quantity > 1) {
+                wc_add_notice('Você só pode ter até 1 assinatura do mesmo produto no seu carrinho.', 'error');
+                return false;
+            }
+        }
+
+        return $passed;
+    }
+
+    public function get_subscription_count($product_id)
+    {
+        $cart = WC()->cart->get_cart();
+        $subscription_count = 0;
+
+        foreach ($cart as $cart_item) {
+            if ($cart_item['data']->get_id() === $product_id) {
+                $subscription_count += $cart_item['quantity'];
+            }
+        }
+
+        return $subscription_count;
+    }
+
+    public function limit_duplicate_subscriptions_cart_update($passed, $cart_item_key, $values, $quantity)
+    {
+        $product_id = $values['product_id'];
+        $product = wc_get_product($product_id);
+
+        if ($this->is_virtual_product($product)) {
+            return $passed;
+        }
+
+        if ($this->subscription_exceeds_limit($product_id, $quantity)) {
+            return false;
+        }
+
+        return $passed;
+    }
+
+    public function is_virtual_product($product)
+    {
+        return $product->is_virtual();
+    }
+
+    public function subscription_exceeds_limit($product_id, $quantity)
+    {
+        if (WC_Subscriptions_Product::is_subscription($product_id)) {
+            $subscription_count = $this->count_subscriptions_in_cart($product_id);
+
+            if ($subscription_count >= 1 && $quantity > 1) {
+                $message = 'Você só pode ter até 1 assinatura do mesmo produto no seu carrinho.';
+                wc_add_notice(__($message, 'vindi-payment-gateway'), 'error');
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    public function count_subscriptions_in_cart($product_id)
+    {
+        $subscription_count = 0;
+        foreach (WC()->cart->get_cart() as $cart_item) {
+            if ($cart_item['data']->get_id() === $product_id) {
+                $subscription_count++;
+            }
+        }
+        return $subscription_count;
+    }
+
+    public function disallow_subscription_single_product_cart($passed, $product_id, $quantity)
+    {
+        $product = wc_get_product($product_id);
+
+        if ($product->is_virtual()) {
+            return $passed;
+        }
+
+        if ($this->is_cart_mixed_with_subscription($product_id)) {
+            wc_add_notice(__('Olá! Finalize a compra da assinatura adicionada
+             ao carrinho antes de adicionar outra assinatura ou produto.', 'vindi-payment-gateway'), 'error');
+            return false;
+        }
+
+        return $passed;
+    }
+
+    public function is_cart_mixed_with_subscription($product_id)
+    {
+        $cart = WC()->cart->get_cart();
+        if (empty($cart)) {
+            return false;
+        }
+
+        $is_subscription = false;
+        $new_product_subscription = WC_Subscriptions_Product::is_subscription($product_id);
+
+        foreach ($cart as $cart_item) {
+            if (WC_Subscriptions_Product::is_subscription($cart_item['data']->get_id())) {
+                $is_subscription = true;
+                break;
+            }
+        }
+
+        return $is_subscription !== $new_product_subscription;
+    }
+}
